Title: 자연수 n개의 합, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240715/자연수 n개의 합/sum-of-n-natural-numbers.cpp
@@ -2,7 +2,7 @@
 #include <limits.h>
 using namespace std;
 
-int s;
+long long s;
 
 int main() {
     cin >> s;
